+++ src/main/java/de/dddns/kirbylink/keepachangelogupdater/model/changelog/category/CategoryDeprecated.java
@@ -1,0 +1,12 @@
+package de.dddns.kirbylink.keepachangelogupdater.model.changelog.category;
+
+import lombok.Builder.Default;
+import lombok.Getter;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+public class CategoryDeprecated extends Category {
+  @Default
+  private CategoryType type = CategoryType.DEPRECATED;
+}